StoreProjectJob relocated
